+++ lithium_ion/electrolytes/LiPF6_Mohtat2020/electrolyte_conductivity_PeymanMPM.py
@@ -1,0 +1,33 @@
+from pybamm import exp, constants
+
+
+def electrolyte_conductivity_PeymanMPM(c_e, T):
+    """
+    Conductivity of LiPF6 in EC:DMC as a function of ion concentration. The original
+    data is from [1]. The fit is from Dualfoil [2].
+
+    References
+    ----------
+    .. [1] C Capiglia et al. 7Li and 19F diffusion coefficients and thermal
+    properties of non-aqueous electrolyte solutions for rechargeable lithium batteries.
+    Journal of power sources 81 (1999): 859-862.
+    .. [2] http://www.cchem.berkeley.edu/jsngrp/fortran.html
+    Parameters
+    ----------
+    c_e: :class:`pybamm.Symbol`
+        Dimensional electrolyte concentration
+    T: :class:`pybamm.Symbol`
+        Dimensional temperature
+
+
+    Returns
+    -------
+    :class:`pybamm.Symbol`
+        Electrolyte conductivity
+    """
+
+    sigma_e = 1.3
+    E_k_e = 34700
+    arrhenius = exp(E_k_e / constants.R * (1 / 298.15 - 1 / T))
+
+    return sigma_e * arrhenius