+++ lithium_ion/electrolytes/lipf6_Ecker2015/electrolyte_conductivity_Ecker2015.py
@@ -1,0 +1,44 @@
+from pybamm import exp, constants
+
+
+def electrolyte_conductivity_Ecker2015(c_e, T):
+    """
+    Conductivity of LiPF6 in EC:DMC as a function of ion concentration [1, 2, 3].
+
+    References
+    ----------
+    .. [1] Ecker, Madeleine, et al. "Parameterization of a physico-chemical model of
+    a lithium-ion battery i. determination of parameters." Journal of the
+    Electrochemical Society 162.9 (2015): A1836-A1848.
+    .. [2] Ecker, Madeleine, et al. "Parameterization of a physico-chemical model of
+    a lithium-ion battery ii. model validation." Journal of The Electrochemical
+    Society 162.9 (2015): A1849-A1857.
+    .. [3] Richardson, Giles, et. al. "Generalised single particle models for
+    high-rate operation of graded lithium-ion electrodes: Systematic derivation
+    and validation." Electrochemica Acta 339 (2020): 135862
+
+    Parameters
+    ----------
+    c_e: :class:`pybamm.Symbol`
+        Dimensional electrolyte concentration
+    T: :class:`pybamm.Symbol`
+        Dimensional temperature
+
+    Returns
+    -------
+    :class:`pybamm.Symbol`
+        Solid diffusivity
+    """
+
+    # mol/m^3 to mol/l
+    cm = 1e-3 * c_e
+
+    # value at T = 296K
+    sigma_e_296 = 0.2667 * cm ** 3 - 1.2983 * cm ** 2 + 1.7919 * cm + 0.1726
+
+    # add temperature dependence
+    E_k_e = 1.71e4
+    C = 296 * exp(E_k_e / (constants.R * 296))
+    sigma_e = C * sigma_e_296 * exp(-E_k_e / (constants.R * T)) / T
+
+    return sigma_e