+++ lithium_ion/positive_electrodes/lico2_Marquis2019/lico2_electrolyte_exchange_current_density_Dualfoil1998.py
@@ -1,0 +1,35 @@
+from pybamm import exp, constants, Parameter
+
+
+def lico2_electrolyte_exchange_current_density_Dualfoil1998(c_e, c_s_surf, T):
+    """
+    Exchange-current density for Butler-Volmer reactions between lico2 and LiPF6 in
+    EC:DMC.
+
+    References
+    ----------
+    .. [2] http://www.cchem.berkeley.edu/jsngrp/fortran.html
+
+    Parameters
+    ----------
+    c_e : :class:`pybamm.Symbol`
+        Electrolyte concentration [mol.m-3]
+    c_s_surf : :class:`pybamm.Symbol`
+        Particle concentration [mol.m-3]
+    T : :class:`pybamm.Symbol`
+        Temperature [K]
+
+    Returns
+    -------
+    :class:`pybamm.Symbol`
+        Exchange-current density [A.m-2]
+    """
+    m_ref = 6 * 10 ** (-7)  # (A/m2)(mol/m3)**1.5 - includes ref concentrations
+    E_r = 39570
+    arrhenius = exp(E_r / constants.R * (1 / 298.15 - 1 / T))
+
+    c_p_max = Parameter("Maximum concentration in positive electrode [mol.m-3]")
+
+    return (
+        m_ref * arrhenius * c_e ** 0.5 * c_s_surf ** 0.5 * (c_p_max - c_s_surf) ** 0.5
+    )