+++ pybamm/models/full_battery_models/lithium_ion/base_lithium_ion_model.py
@@ -44,6 +44,36 @@ class BaseModel(pybamm.BaseBatteryModel):
             )
         self.set_standard_output_variables()
 
+    @property
+    def default_parameter_values(self):
+        if self.half_cell:
+            return pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Xu2019)
+        else:
+            return pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Marquis2019)
+
+    @property
+    def default_quick_plot_variables(self):
+        if self.half_cell:
+            return [
+                "Electrolyte concentration [mol.m-3]",
+                "Positive particle surface concentration [mol.m-3]",
+                "Current [A]",
+                "Electrolyte potential [V]",
+                "Positive electrode potential [V]",
+                "Terminal voltage [V]",
+            ]
+        else:
+            return [
+                "Negative particle surface concentration [mol.m-3]",
+                "Electrolyte concentration [mol.m-3]",
+                "Positive particle surface concentration [mol.m-3]",
+                "Current [A]",
+                "Negative electrode potential [V]",
+                "Electrolyte potential [V]",
+                "Positive electrode potential [V]",
+                "Terminal voltage [V]",
+            ]
+
     def set_standard_output_variables(self):
         super().set_standard_output_variables()
 
@@ -118,53 +148,86 @@ class BaseModel(pybamm.BaseBatteryModel):
             }
         )
 
-    def set_sei_submodel(self):
+    def set_summary_variables(self):
+        """
+        Sets the default summary variables.
+        """
+        summary_variables = [
+            "Positive electrode capacity [A.h]",
+            # LAM, LLI
+            "Loss of active material in positive electrode [%]",
+            "Loss of lithium inventory [%]",
+            "Loss of lithium inventory, including electrolyte [%]",
+            # Total lithium
+            "Total lithium [mol]",
+            "Total lithium in electrolyte [mol]",
+            "Total lithium in positive electrode [mol]",
+            "Total lithium in particles [mol]",
+            # Lithium lost
+            "Total lithium lost [mol]",
+            "Total lithium lost from particles [mol]",
+            "Total lithium lost from electrolyte [mol]",
+            "Loss of lithium to SEI [mol]",
+            "Loss of lithium to lithium plating [mol]",
+            "Loss of capacity to SEI [A.h]",
+            "Loss of capacity to lithium plating [A.h]",
+            "Total lithium lost to side reactions [mol]",
+            "Total capacity lost to side reactions [A.h]",
+            # Resistance
+            "Local ECM resistance [Ohm]",
+        ]
 
-        # SEI
+        if not self.half_cell:
+            summary_variables += [
+                "Negative electrode capacity [A.h]",
+                "Loss of active material in negative electrode [%]",
+                "Total lithium in negative electrode [mol]",
+            ]
+
+        self.summary_variables = summary_variables
+
+    def set_sei_submodel(self):
+        if self.half_cell:
+            reaction_loc = "interface"
+        elif self.x_average:
+            reaction_loc = "x-average"
+        else:
+            reaction_loc = "full electrode"
+
         if self.options["SEI"] == "none":
             self.submodels["sei"] = pybamm.sei.NoSEI(self.param, self.options)
-
-        if self.options["SEI"] == "constant":
-            self.submodels["sei"] = pybamm.sei.ConstantSEI(self.param)
-
+        elif self.options["SEI"] == "constant":
+            self.submodels["sei"] = pybamm.sei.ConstantSEI(self.param, self.options)
         elif self.options["SEI"] == "reaction limited":
             self.submodels["sei"] = pybamm.sei.ReactionLimited(
-                self.param, self.x_average
+                self.param, reaction_loc, self.options
             )
-
         elif self.options["SEI"] == "solvent-diffusion limited":
             self.submodels["sei"] = pybamm.sei.SolventDiffusionLimited(
-                self.param, self.x_average
+                self.param, reaction_loc, self.options
             )
-
         elif self.options["SEI"] == "electron-migration limited":
             self.submodels["sei"] = pybamm.sei.ElectronMigrationLimited(
-                self.param, self.x_average
+                self.param, reaction_loc, self.options
             )
-
         elif self.options["SEI"] == "interstitial-diffusion limited":
             self.submodels["sei"] = pybamm.sei.InterstitialDiffusionLimited(
-                self.param, self.x_average
+                self.param, reaction_loc, self.options
             )
-
         elif self.options["SEI"] == "ec reaction limited":
             self.submodels["sei"] = pybamm.sei.EcReactionLimited(
-                self.param, self.x_average
+                self.param, reaction_loc, self.options
             )
 
     def set_lithium_plating_submodel(self):
-
-        # negative electrode
         if self.options["lithium plating"] == "none":
             self.submodels["lithium plating"] = pybamm.lithium_plating.NoPlating(
                 self.param, self.options
             )
-
         elif self.options["lithium plating"] == "reversible":
             self.submodels[
                 "lithium plating"
             ] = pybamm.lithium_plating.ReversiblePlating(self.param, self.x_average)
-
         elif self.options["lithium plating"] == "irreversible":
             self.submodels[
                 "lithium plating"