+++ pybamm/models/submodels/interface/sei/electron_migration_limited.py
@@ -13,25 +13,31 @@ class ElectronMigrationLimited(BaseModel):
     ----------
     param : parameter class
         The parameters to use for this submodel
-    x_average : bool
-        Whether to use x-averaged variables (SPM, SPMe, etc) or full variables (DFN)
+    reaction_loc : str
+        Where the reaction happens: "x-average" (SPM, SPMe, etc),
+        "full electrode" (full DFN), or "interface" (half-cell DFN)
+    options : dict, optional
+        A dictionary of options to be passed to the model.
 
     **Extends:** :class:`pybamm.sei.BaseModel`
     """
 
-    def __init__(self, param, x_average):
-        super().__init__(param)
-        self.x_average = x_average
+    def __init__(self, param, reaction_loc, options=None):
+        super().__init__(param, options=options)
+        self.reaction_loc = reaction_loc
 
     def get_fundamental_variables(self):
-        if self.x_average is True:
+        if self.reaction_loc == "x-average":
             L_inner_av = pybamm.standard_variables.L_inner_av
             L_outer_av = pybamm.standard_variables.L_outer_av
             L_inner = pybamm.PrimaryBroadcast(L_inner_av, "negative electrode")
             L_outer = pybamm.PrimaryBroadcast(L_outer_av, "negative electrode")
-        else:
+        elif self.reaction_loc == "full electrode":
             L_inner = pybamm.standard_variables.L_inner
             L_outer = pybamm.standard_variables.L_outer
+        elif self.reaction_loc == "interface":
+            L_inner = pybamm.standard_variables.L_inner_interface
+            L_outer = pybamm.standard_variables.L_outer_interface
 
         variables = self._get_standard_thickness_variables(L_inner, L_outer)
         variables.update(self._get_standard_concentration_variables(variables))
@@ -40,7 +46,10 @@ class ElectronMigrationLimited(BaseModel):
 
     def get_coupled_variables(self, variables):
         L_sei_inner = variables["Inner SEI thickness"]
-        phi_s_n = variables["Negative electrode potential"]
+        if self.reaction_loc == "interface":
+            phi_s_n = variables["Lithium metal interface electrode potential"]
+        else:
+            phi_s_n = variables["Negative electrode potential"]
 
         U_inner = self.param.U_inner_electron
         C_sei = self.param.C_sei_electron
@@ -59,7 +68,7 @@ class ElectronMigrationLimited(BaseModel):
         return variables
 
     def set_rhs(self, variables):
-        if self.x_average is True:
+        if self.reaction_loc == "x-average":
             L_inner = variables["X-averaged inner SEI thickness"]
             L_outer = variables["X-averaged outer SEI thickness"]
             j_inner = variables["X-averaged inner SEI interfacial current density"]
@@ -79,7 +88,7 @@ class ElectronMigrationLimited(BaseModel):
         }
 
     def set_initial_conditions(self, variables):
-        if self.x_average is True:
+        if self.reaction_loc == "x-average":
             L_inner = variables["X-averaged inner SEI thickness"]
             L_outer = variables["X-averaged outer SEI thickness"]
         else: