+++ pybamm/solvers/processed_variable.py
@@ -545,9 +545,10 @@ class ProcessedVariable(object):
                 self.initialise_sensitivity_explicit_forward()
             else:
                 raise ValueError(
-                    "Cannot compute sensitivities. The 'sensitivities' argument of the "
-                    "solver.solve should be changed from 'None' to allow sensitivities "
-                    "calculations. Check solver documentation for details."
+                    "Cannot compute sensitivities. The 'calculate_sensitivities' "
+                    "argument of the solver.solve should be changed from 'None' to "
+                    "allow sensitivities calculations. Check solver documentation for "
+                    "details."
                 )
         return self._sensitivities