+++ tests/unit/test_batch_study.py
@@ -1,42 +1,32 @@
 """
 Tests for the batch_study.py
 """
+import os
 import pybamm
 import unittest
 
 spm = pybamm.lithium_ion.SPM()
-dfn = pybamm.lithium_ion.DFN()
+spm_uniform = pybamm.lithium_ion.SPM({"particle": "uniform profile"})
 casadi_safe = pybamm.CasadiSolver(mode="safe")
 casadi_fast = pybamm.CasadiSolver(mode="fast")
-cccv = pybamm.Experiment(
-    [
-        (
-            "Discharge at C/5 for 10 hours or until 3.3 V",
-            "Rest for 1 hour",
-            "Charge at 1 A until 4.1 V",
-            "Hold at 4.1 V until 10 mA",
-            "Rest for 1 hour",
-        ),
-    ]
-    * 3
-)
-gitt = pybamm.Experiment(
-    [("Discharge at C/20 for 1 hour", "Rest for 1 hour")] * 20,
-)
+exp1 = pybamm.Experiment([("Discharge at C/5 for 10 minutes", "Rest for 1 hour")])
+exp2 = pybamm.Experiment([("Discharge at C/20 for 10 minutes", "Rest for 1 hour")])
 
-bs_false_only_models = pybamm.BatchStudy(models={"SPM": spm, "DFN": dfn})
+bs_false_only_models = pybamm.BatchStudy(
+    models={"SPM": spm, "SPM uniform": spm_uniform}
+)
 bs_true_only_models = pybamm.BatchStudy(
-    models={"SPM": spm, "DFN": dfn}, permutations=True
+    models={"SPM": spm, "SPM uniform": spm_uniform}, permutations=True
 )
 bs_false = pybamm.BatchStudy(
-    models={"SPM": spm, "DFN": dfn},
+    models={"SPM": spm, "SPM uniform": spm_uniform},
     solvers={"casadi safe": casadi_safe, "casadi fast": casadi_fast},
-    experiments={"cccv": cccv, "gitt": gitt},
+    experiments={"exp1": exp1, "exp2": exp2},
 )
 bs_true = pybamm.BatchStudy(
-    models={"SPM": spm, "DFN": dfn},
+    models={"SPM": spm, "SPM uniform": spm_uniform},
     solvers={"casadi safe": casadi_safe, "casadi fast": casadi_fast},
-    experiments={"gitt": gitt},
+    experiments={"exp2": exp2},
     permutations=True,
 )
 
@@ -47,8 +37,7 @@ class TestBatchStudy(unittest.TestCase):
         for name in pybamm.BatchStudy.INPUT_LIST:
             with self.assertRaises(ValueError):
                 pybamm.BatchStudy(
-                    models={"SPM": spm, "DFN": dfn},
-                    **{name: {None}}
+                    models={"SPM": spm, "SPM uniform": spm_uniform}, **{name: {None}}
                 )
 
         # Tests for None when only models are given with permutations=False
@@ -61,6 +50,7 @@ class TestBatchStudy(unittest.TestCase):
 
         # Tests for BatchStudy when permutations=False
         bs_false.solve()
+        bs_false.plot(testing=True)
         self.assertEqual(2, len(bs_false.sims))
         for num in range(len(bs_false.sims)):
             output_model = bs_false.sims[num].model.name
@@ -68,9 +58,7 @@ class TestBatchStudy(unittest.TestCase):
             self.assertIn(output_model, models_list)
 
             output_solver = bs_false.sims[num].solver.name
-            solvers_list = [
-                solver.name for solver in bs_false.solvers.values()
-            ]
+            solvers_list = [solver.name for solver in bs_false.solvers.values()]
             self.assertIn(output_solver, solvers_list)
 
             output_experiment = bs_false.sims[
@@ -84,6 +72,7 @@ class TestBatchStudy(unittest.TestCase):
 
         # Tests for BatchStudy when permutations=True
         bs_true.solve()
+        bs_true.plot(testing=True)
         self.assertEqual(4, len(bs_true.sims))
         for num in range(len(bs_true.sims)):
             output_model = bs_true.sims[num].model.name
@@ -103,24 +92,18 @@ class TestBatchStudy(unittest.TestCase):
             ]
             self.assertIn(output_experiment, experiments_list)
 
-    def test_plot(self):
-        # Tests for BatchStudy when permutations=False
-        bs_false.solve()
-        bs_false.plot(testing=True)
-        self.assertEqual(2, len(bs_false.sims))
-        for num in range(len(bs_false.sims)):
-            output_model = bs_false.sims[num].all_models[0].name
-            models_list = [model.name for model in bs_false.models.values()]
-            self.assertIn(output_model, models_list)
+    def test_create_gif(self):
+        bs = pybamm.BatchStudy({"spm": pybamm.lithium_ion.SPM()})
+        bs.solve([0, 10])
 
-        # Tests for BatchStudy when permutations=True
-        bs_true.solve()
-        bs_true.plot(testing=True)
-        self.assertEqual(4, len(bs_true.sims))
-        for num in range(len(bs_true.sims)):
-            output_model = bs_true.sims[num].all_models[0].name
-            models_list = [model.name for model in bs_true.models.values()]
-            self.assertIn(output_model, models_list)
+        # create a GIF before calling the plot method
+        bs.create_gif(number_of_images=3, duration=1)
+
+        # create a GIF after calling the plot method
+        bs.plot(testing=True)
+        bs.create_gif(number_of_images=3, duration=1)
+
+        os.remove("plot.gif")
 
 
 if __name__ == "__main__":