+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_base_lithium_ion_model.py
@@ -3,6 +3,7 @@
 #
 import pybamm
 import unittest
+import os
 
 
 class TestBaseLithiumIonModel(unittest.TestCase):
@@ -13,6 +14,23 @@ class TestBaseLithiumIonModel(unittest.TestCase):
             pybamm.lithium_ion.BaseModel(
                 {"cell geometry": "arbitrary", "thermal": "x-lumped"}
             )
+
+    def test_default_parameters(self):
+        # check parameters are read in ok
+        model = pybamm.lithium_ion.BaseModel()
+        self.assertEqual(
+            model.default_parameter_values["Reference temperature [K]"], 298.15
+        )
+
+        # change path and try again
+
+        cwd = os.getcwd()
+        os.chdir("..")
+        model = pybamm.lithium_ion.BaseModel()
+        self.assertEqual(
+            model.default_parameter_values["Reference temperature [K]"], 298.15
+        )
+        os.chdir(cwd)
 
 
 if __name__ == "__main__":