+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_dfn_half_cell.py
@@ -13,6 +13,41 @@ class TestDFNHalfCell(unittest.TestCase):
         model.check_well_posedness()
 
 
+class TestDFNHalfCellWithSEI(unittest.TestCase):
+    def test_well_posed_constant(self):
+        options = {"working electrode": "positive", "SEI": "constant"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+    def test_well_posed_reaction_limited(self):
+        options = {"working electrode": "positive", "SEI": "reaction limited"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+    def test_well_posed_solvent_diffusion_limited(self):
+        options = {"working electrode": "positive", "SEI": "solvent-diffusion limited"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+    def test_well_posed_electron_migration_limited(self):
+        options = {"working electrode": "positive", "SEI": "electron-migration limited"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+    def test_well_posed_interstitial_diffusion_limited(self):
+        options = {
+            "working electrode": "positive",
+            "SEI": "interstitial-diffusion limited",
+        }
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+    def test_well_posed_ec_reaction_limited(self):
+        options = {"working electrode": "positive", "SEI": "ec reaction limited"}
+        model = pybamm.lithium_ion.DFN(options)
+        model.check_well_posedness()
+
+
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys